refactor: flatten enqueue for all func
